Add files via upload
1.  Create a Sender Email Account
Create a new email account for sending emails with the ESP32. If you want to use a Gmail account, go to this link to create a new one.
2.  Create an App Password
You need to create an app password so that the ESP32 is able to send emails using your Gmail account. An App Password is a 16-digit passcode that gives a less secure app or device permission to access your Google Account. Learn more about sign-in with app passwords here.

An app password can only be used with accounts that have 2-step verification turned on.

Open your Google Account.
In the navigation panel, select Security.
Under “Signing in to Google,” select 2-Step Verification > Get started.
Follow the on-screen steps.
After enabling 2-step verification, you can create an app password.

Open your Google Account.
In the search panel, search for App Passwords.
Open the App Passwords menu.